[MOD] #267 BaseResponse 사용하도록 dto 변경
